backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.9
pydantic>=2.10.0
pydantic-settings>=2.6.0
openai>=1.57.0
aiofiles>=24.1.0
pymupdf>=1.24.0
pypdf>=5.0.0
Pillow>=10.4.0
pdf2image>=1.17.0
pdf2docx>=0.5.8
python-docx>=1.1.0
python-pptx>=1.0.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
xhtml2pdf>=0.2.16
pikepdf>=9.0.0
pytesseract>=0.3.13
python-magic-bin>=0.4.14; sys_platform == "win32"
filetype>=1.2.0
httpx>=0.28.0
google-generativeai>=0.8.0
motor>=3.0.0
pymongo>=4.0.0
certifi>=2024.7.4
cloudinary>=1.41.0
